Update fedexrest.php
Added customer friendly description for the checkout and estimator areas..

Update fedexrest.php

Adding the actual part that keeps the module's full name in the backend.

Update fedexrest.php

Changing `$this->description` to have a customer friendly label. Will see "FedEx" as the method instead of "FedEx REST API" on the outside.